Create readme
This is a website scraper I created to assist with job hunting and practice coding in python. I initially used bs4 only to load indeed.com, but needed to use selenium to open the dynamic webpage first. I then passed the  open website as an object into beautifulsoup4 and then extract information from appropriate tags. 

In the future, I would like to scale my code to deal with pagination, and store information in either JSON dictionary or PANDAS to allow for integrate analysis. 

Thank you for your attention